Release 1.0
End of our journey with MinMax Algorithm

After a week of hard work, we made a game fully functional (i hope so...), with options to play with AI or with a friend.

We started with no knowledge about how to make a game in godot, after this journey we can say that we know.... like 25% of the power of godot game engine.

Making this game was fun after all !

Co-Authored-By: Wiktoria Zaborowska <115507011+WiktoriaZab@users.noreply.github.com>